requirements.txt
langchain
evadb
pandas
numpy
openai
python-dotenv
pymupdf